+++ lib/puzzle_cube.dart
@@ -8,7 +8,7 @@ export 'src/models/cube_color_validator.dart';
 export 'src/models/cube_colors.dart';
 export 'src/models/cube_move.dart';
 export 'src/models/cubie_model.dart';
-export 'src/models/rubiks_cube_state.dart';
+export 'src/models/puzzle_cube_state.dart';
 export 'src/rendering/cube_hit_test.dart';
 export 'src/rendering/cubie_builder.dart';
 export 'src/widgets/cube_controller.dart' show Cube, CubeController;